🔵 Move DefaultDiscountService to domain
